Simplify internal frame setup functions
This was overly complicated for no good reason.
